Move eudemo blanket builder module
